requirements.txt
pandas
flask
transformers
spacy_udpipe
python-dotenv
Flask-CORS
